fix: bypass node-lief for ELF binaries in nativeInstallationLoader
- Extract ELF-only code into elfInstallation.ts (no LIEF dependency)
- nativeInstallationLoader.ts now bypasses LIEF gate for ELF files,
  fixing tweakcc --apply on NixOS where node-lief cannot load
- Loosen empty-settings guard in --apply mode so version badge applies
  even when settings = {}
- nix/patch.nix: add set -e for loud failures, fix config.json shape
  to wrap settings under {"settings": {...}} as readConfigFile expects
