+++ src/nativeInstallation.ts
@@ -6,6 +6,23 @@ import fs from 'node:fs';
 import { execSync } from 'node:child_process';
 import LIEF from 'node-lief';
 import { isDebug, debug } from './utils';
+import {
+  BUN_TRAILER,
+  SIZEOF_OFFSETS,
+  type BunOffsets,
+  type BunData,
+  getStringPointerContent,
+  isClaudeModule,
+  detectModuleStructSize,
+  mapModules,
+  parseOffsets,
+  extractBunDataFromSection,
+  isELFFile,
+  extractBunDataFromELFRaw,
+  rebuildBunData,
+  buildSectionData,
+  repackELFRaw,
+} from './elfInstallation';
 
 // ============================================================================
 // Nix binary wrapper detection
@@ -134,438 +151,6 @@ export function resolveNixBinaryWrapper(binaryPath: string): string | null {
   } catch (error) {
     debug('resolveNixBinaryWrapper: error during detection:', error);
     return null;
-  }
-}
-
-/**
- * Constants for Bun trailer and serialized layout sizes.
- *
- * Bun data layout (normalized across formats) is:
- * [data...][OFFSETS struct][BUN_TRAILER]
- *
- * Where OFFSETS struct (SIZEOF_OFFSETS bytes) is:
- * - byteCount:   u64  (total size of [data][OFFSETS][BUN_TRAILER])
- * - modulesPtr:  { u32 offset, u32 length } into [data...] for modules table
- * - entryPointId: u32
- * - compileExecArgvPtr: { u32 offset, u32 length }
- * - flags: u32
- */
-const BUN_TRAILER = Buffer.from('\n---- Bun! ----\n');
-
-// Size constants for binary structures
-const SIZEOF_OFFSETS = 32;
-const SIZEOF_STRING_POINTER = 8;
-// Module struct sizes vary by Bun version:
-// - Old format (pre-ESM bytecode, before Bun ~1.3.7): 4 StringPointers + 4 u8s = 36 bytes
-// - New format (ESM bytecode, Bun ~1.3.7+): 6 StringPointers + 4 u8s = 52 bytes
-const SIZEOF_MODULE_OLD = 4 * SIZEOF_STRING_POINTER + 4;
-const SIZEOF_MODULE_NEW = 6 * SIZEOF_STRING_POINTER + 4;
-
-// Types
-interface StringPointer {
-  offset: number;
-  length: number;
-}
-
-interface BunOffsets {
-  byteCount: bigint | number;
-  modulesPtr: StringPointer;
-  entryPointId: number;
-  compileExecArgvPtr: StringPointer;
-  flags: number;
-}
-
-interface BunModule {
-  name: StringPointer;
-  contents: StringPointer;
-  sourcemap: StringPointer;
-  bytecode: StringPointer;
-  moduleInfo: StringPointer;
-  bytecodeOriginPath: StringPointer;
-  encoding: number;
-  loader: number;
-  moduleFormat: number;
-  side: number;
-}
-
-interface BunData {
-  bunOffsets: BunOffsets;
-  bunData: Buffer;
-  /** Header size used in section format: 4 for old format (Bun < 1.3.4), 8 for new format. Only for Mach-O and PE. */
-  sectionHeaderSize?: number;
-  /** Detected module struct size: SIZEOF_MODULE_OLD (36) or SIZEOF_MODULE_NEW (52). */
-  moduleStructSize: number;
-}
-
-/**
- * Read a StringPointer slice from given buffer.
- */
-function getStringPointerContent(
-  buffer: Buffer,
-  stringPointer: StringPointer
-): Buffer {
-  return buffer.subarray(
-    stringPointer.offset,
-    stringPointer.offset + stringPointer.length
-  );
-}
-
-function parseStringPointer(buffer: Buffer, offset: number): StringPointer {
-  return {
-    offset: buffer.readUInt32LE(offset),
-    length: buffer.readUInt32LE(offset + 4),
-  };
-}
-
-/**
- * True if the module represents the native claude entrypoint.
- */
-function isClaudeModule(moduleName: string): boolean {
-  return (
-    moduleName.endsWith('/claude') ||
-    moduleName === 'claude' ||
-    moduleName.endsWith('/claude.exe') ||
-    moduleName === 'claude.exe' ||
-    moduleName.endsWith('/src/entrypoints/cli.js') ||
-    moduleName === 'src/entrypoints/cli.js'
-  );
-}
-
-/**
- * Detects the module struct size from the modules list byte length.
- * Returns SIZEOF_MODULE_NEW (52) or SIZEOF_MODULE_OLD (36).
- */
-function detectModuleStructSize(modulesListLength: number): number {
-  const fitsNew = modulesListLength % SIZEOF_MODULE_NEW === 0;
-  const fitsOld = modulesListLength % SIZEOF_MODULE_OLD === 0;
-
-  if (fitsNew && !fitsOld) return SIZEOF_MODULE_NEW;
-  if (fitsOld && !fitsNew) return SIZEOF_MODULE_OLD;
-  if (fitsNew && fitsOld) {
-    // Ambiguous — prefer new format (more likely with recent Bun versions)
-    debug(
-      `detectModuleStructSize: Ambiguous module list length ${modulesListLength}, assuming new format`
-    );
-    return SIZEOF_MODULE_NEW;
-  }
-
-  // Neither fits cleanly — try new format as default
-  debug(
-    `detectModuleStructSize: Module list length ${modulesListLength} doesn't cleanly divide by either struct size, assuming new format`
-  );
-  return SIZEOF_MODULE_NEW;
-}
-
-/**
- * Iterates over modules in the Bun data and calls visitor for each.
- * Handles all module parsing and iteration logic in one place.
- */
-function mapModules<T>(
-  bunData: Buffer,
-  bunOffsets: BunOffsets,
-  moduleStructSize: number,
-  visitor: (
-    module: BunModule,
-    moduleName: string,
-    index: number
-  ) => T | undefined
-): T | undefined {
-  const modulesListBytes = getStringPointerContent(
-    bunData,
-    bunOffsets.modulesPtr
-  );
-  const modulesListCount = Math.floor(
-    modulesListBytes.length / moduleStructSize
-  );
-
-  for (let i = 0; i < modulesListCount; i++) {
-    const offset = i * moduleStructSize;
-    const module = parseCompiledModuleGraphFile(
-      modulesListBytes,
-      offset,
-      moduleStructSize
-    );
-    const moduleName = getStringPointerContent(bunData, module.name).toString(
-      'utf-8'
-    );
-
-    const result = visitor(module, moduleName, i);
-    if (result !== undefined) {
-      return result;
-    }
-  }
-
-  return undefined;
-}
-
-function parseOffsets(buffer: Buffer): BunOffsets {
-  let pos = 0;
-  const byteCount = buffer.readBigUInt64LE(pos);
-  pos += 8;
-  const modulesPtr = parseStringPointer(buffer, pos);
-  pos += 8;
-  const entryPointId = buffer.readUInt32LE(pos);
-  pos += 4;
-  const compileExecArgvPtr = parseStringPointer(buffer, pos);
-  pos += 8;
-  const flags = buffer.readUInt32LE(pos);
-
-  return { byteCount, modulesPtr, entryPointId, compileExecArgvPtr, flags };
-}
-
-function parseCompiledModuleGraphFile(
-  buffer: Buffer,
-  offset: number,
-  moduleStructSize: number
-): BunModule {
-  let pos = offset;
-  const name = parseStringPointer(buffer, pos);
-  pos += 8;
-  const contents = parseStringPointer(buffer, pos);
-  pos += 8;
-  const sourcemap = parseStringPointer(buffer, pos);
-  pos += 8;
-  const bytecode = parseStringPointer(buffer, pos);
-  pos += 8;
-
-  let moduleInfo: StringPointer;
-  let bytecodeOriginPath: StringPointer;
-  if (moduleStructSize === SIZEOF_MODULE_NEW) {
-    moduleInfo = parseStringPointer(buffer, pos);
-    pos += 8;
-    bytecodeOriginPath = parseStringPointer(buffer, pos);
-    pos += 8;
-  } else {
-    moduleInfo = { offset: 0, length: 0 };
-    bytecodeOriginPath = { offset: 0, length: 0 };
-  }
-
-  const encoding = buffer.readUInt8(pos);
-  pos += 1;
-  const loader = buffer.readUInt8(pos);
-  pos += 1;
-  const moduleFormat = buffer.readUInt8(pos);
-  pos += 1;
-  const side = buffer.readUInt8(pos);
-
-  return {
-    name,
-    contents,
-    sourcemap,
-    bytecode,
-    moduleInfo,
-    bytecodeOriginPath,
-    encoding,
-    loader,
-    moduleFormat,
-    side,
-  };
-}
-
-/**
- * Parses Bun data blob that contains: [data][offsets][trailer]
- * This is the common structure across all formats after extraction.
- */
-function parseBunDataBlob(bunDataContent: Buffer): {
-  bunOffsets: BunOffsets;
-  bunData: Buffer;
-  moduleStructSize: number;
-} {
-  if (bunDataContent.length < SIZEOF_OFFSETS + BUN_TRAILER.length) {
-    throw new Error('BUN data is too small to contain trailer and offsets');
-  }
-
-  // Verify trailer
-  const trailerStart = bunDataContent.length - BUN_TRAILER.length;
-  const trailerBytes = bunDataContent.subarray(trailerStart);
-
-  debug(`parseBunDataBlob: Expected trailer: ${BUN_TRAILER.toString('hex')}`);
-  debug(`parseBunDataBlob: Got trailer: ${trailerBytes.toString('hex')}`);
-
-  if (!trailerBytes.equals(BUN_TRAILER)) {
-    debug(`Expected: ${BUN_TRAILER.toString('hex')}`);
-    debug(`Got: ${trailerBytes.toString('hex')}`);
-    throw new Error('BUN trailer bytes do not match trailer');
-  }
-
-  // Parse Offsets structure
-  const offsetsStart =
-    bunDataContent.length - SIZEOF_OFFSETS - BUN_TRAILER.length;
-  const offsetsBytes = bunDataContent.subarray(
-    offsetsStart,
-    offsetsStart + SIZEOF_OFFSETS
-  );
-  const bunOffsets = parseOffsets(offsetsBytes);
-  const moduleStructSize = detectModuleStructSize(bunOffsets.modulesPtr.length);
-
-  return {
-    bunOffsets,
-    bunData: bunDataContent,
-    moduleStructSize,
-  };
-}
-
-/**
- * Section format helper (for Mach-O and PE):
- * Old format (Bun < 1.3.4): [u32 size][size bytes of Bun data blob...]
- * New format (Bun >= 1.3.4): [u64 size][size bytes of Bun data blob...]
- *
- * Size is the length of the Bun blob (which itself is [data][OFFSETS][TRAILER]).
- * We detect which format by checking if (headerSize + size) matches the section length.
- */
-function extractBunDataFromSection(sectionData: Buffer): BunData {
-  if (sectionData.length < 4) {
-    throw new Error('Section data too small');
-  }
-
-  debug(`extractBunDataFromSection: sectionData.length=${sectionData.length}`);
-
-  // Try u32 header (old format, Bun < 1.3.4)
-  const bunDataSizeU32 = sectionData.readUInt32LE(0);
-  const expectedLengthU32 = 4 + bunDataSizeU32;
-
-  // Try u64 header (new format, Bun >= 1.3.4) - only if we have enough bytes
-  const bunDataSizeU64 =
-    sectionData.length >= 8 ? Number(sectionData.readBigUInt64LE(0)) : 0;
-  const expectedLengthU64 = 8 + bunDataSizeU64;
-
-  debug(
-    `extractBunDataFromSection: u32 header would give size=${bunDataSizeU32}, expected total=${expectedLengthU32}`
-  );
-  debug(
-    `extractBunDataFromSection: u64 header would give size=${bunDataSizeU64}, expected total=${expectedLengthU64}`
-  );
-
-  let headerSize: number;
-  let bunDataSize: number;
-
-  // Check which format matches the section length (allowing for padding up to 4KB)
-  if (
-    sectionData.length >= 8 &&
-    expectedLengthU64 <= sectionData.length &&
-    expectedLengthU64 >= sectionData.length - 4096
-  ) {
-    // u64 format matches
-    headerSize = 8;
-    bunDataSize = bunDataSizeU64;
-    debug(
-      `extractBunDataFromSection: detected u64 header format (Bun >= 1.3.4)`
-    );
-  } else if (
-    expectedLengthU32 <= sectionData.length &&
-    expectedLengthU32 >= sectionData.length - 4096
-  ) {
-    // u32 format matches
-    headerSize = 4;
-    bunDataSize = bunDataSizeU32;
-    debug(
-      `extractBunDataFromSection: detected u32 header format (Bun < 1.3.4)`
-    );
-  } else {
-    throw new Error(
-      `Cannot determine section header format: sectionData.length=${sectionData.length}, ` +
-        `u64 would expect ${expectedLengthU64}, u32 would expect ${expectedLengthU32}`
-    );
-  }
-
-  debug(`extractBunDataFromSection: bunDataSize from header=${bunDataSize}`);
-
-  const bunDataContent = sectionData.subarray(
-    headerSize,
-    headerSize + bunDataSize
-  );
-
-  debug(
-    `extractBunDataFromSection: bunDataContent.length=${bunDataContent.length}`
-  );
-
-  const { bunOffsets, bunData, moduleStructSize } =
-    parseBunDataBlob(bunDataContent);
-
-  return {
-    bunOffsets,
-    bunData,
-    sectionHeaderSize: headerSize,
-    moduleStructSize,
-  };
-}
-
-const ELF_MAGIC = Buffer.from([0x7f, 0x45, 0x4c, 0x46]);
-
-/** Returns true if the file at filePath begins with the ELF magic bytes. */
-function isELFFile(filePath: string): boolean {
-  let fd: number | null = null;
-  try {
-    fd = fs.openSync(filePath, 'r');
-    const buf = Buffer.allocUnsafe(4);
-    const bytesRead = fs.readSync(fd, buf, 0, 4, 0);
-    return bytesRead === 4 && buf.equals(ELF_MAGIC);
-  } catch {
-    return false;
-  } finally {
-    if (fd !== null) fs.closeSync(fd);
-  }
-}
-
-/**
- * Extracts Bun data from an ELF binary by reading the file tail directly,
- * without using LIEF. Uses BunOffsets.byteCount (from the offsets struct) as
- * the authoritative data region size. The trailing u64 footer is not used.
- */
-function extractBunDataFromELFRaw(filePath: string): BunData {
-  const fd = fs.openSync(filePath, 'r');
-  try {
-    const { size: fileSize } = fs.fstatSync(fd);
-    const tailSize = SIZEOF_OFFSETS + BUN_TRAILER.length + 8;
-
-    if (fileSize < tailSize) {
-      throw new Error('File too small to contain Bun data');
-    }
-
-    const tailBuffer = Buffer.allocUnsafe(tailSize);
-    fs.readSync(fd, tailBuffer, 0, tailSize, fileSize - tailSize);
-
-    const trailerStart = tailSize - 8 - BUN_TRAILER.length;
-    const trailerBytes = tailBuffer.subarray(
-      trailerStart,
-      trailerStart + BUN_TRAILER.length
-    );
-    if (!trailerBytes.equals(BUN_TRAILER)) {
-      throw new Error('BUN trailer not found in ELF file');
-    }
-
-    const offsetsBytes = tailBuffer.subarray(0, SIZEOF_OFFSETS);
-    const bunOffsets = parseOffsets(offsetsBytes);
-    const byteCount =
-      typeof bunOffsets.byteCount === 'bigint'
-        ? Number(bunOffsets.byteCount)
-        : bunOffsets.byteCount;
-
-    if (byteCount <= 0 || byteCount >= fileSize) {
-      throw new Error(`ELF byteCount out of range: ${byteCount}`);
-    }
-
-    const dataStart =
-      fileSize - 8 - BUN_TRAILER.length - SIZEOF_OFFSETS - byteCount;
-    if (dataStart < 0) {
-      throw new Error('ELF data region extends before start of file');
-    }
-
-    const dataBuffer = Buffer.allocUnsafe(byteCount);
-    fs.readSync(fd, dataBuffer, 0, byteCount, dataStart);
-
-    const bunDataBlob = Buffer.concat([dataBuffer, offsetsBytes, trailerBytes]);
-    const moduleStructSize = detectModuleStructSize(
-      bunOffsets.modulesPtr.length
-    );
-
-    debug(
-      `extractBunDataFromELFRaw: byteCount=${byteCount}, moduleStructSize=${moduleStructSize}`
-    );
-
-    return { bunOffsets, bunData: bunDataBlob, moduleStructSize };
-  } finally {
-    fs.closeSync(fd);
   }
 }
 
@@ -790,231 +375,6 @@ export function extractClaudeJsFromNativeInstallation(
   }
 }
 
-function rebuildBunData(
-  bunData: Buffer,
-  bunOffsets: BunOffsets,
-  modifiedClaudeJs: Buffer | null,
-  moduleStructSize: number
-): Buffer {
-  // Phase 1: Collect all string data
-  const stringsData: Buffer[] = [];
-  const modulesMetadata: Array<{
-    name: Buffer;
-    contents: Buffer;
-    sourcemap: Buffer;
-    bytecode: Buffer;
-    moduleInfo: Buffer;
-    bytecodeOriginPath: Buffer;
-    encoding: number;
-    loader: number;
-    moduleFormat: number;
-    side: number;
-  }> = [];
-
-  // Use mapModules to iterate and collect module data
-  mapModules(bunData, bunOffsets, moduleStructSize, (module, moduleName) => {
-    const nameBytes = getStringPointerContent(bunData, module.name);
-
-    // Check if this is claude.js and we have modified contents
-    let contentsBytes: Buffer;
-    if (modifiedClaudeJs && isClaudeModule(moduleName)) {
-      contentsBytes = modifiedClaudeJs;
-    } else {
-      contentsBytes = getStringPointerContent(bunData, module.contents);
-    }
-
-    const sourcemapBytes = getStringPointerContent(bunData, module.sourcemap);
-    const bytecodeBytes = getStringPointerContent(bunData, module.bytecode);
-    const moduleInfoBytes = getStringPointerContent(bunData, module.moduleInfo);
-    const bytecodeOriginPathBytes = getStringPointerContent(
-      bunData,
-      module.bytecodeOriginPath
-    );
-
-    modulesMetadata.push({
-      name: nameBytes,
-      contents: contentsBytes,
-      sourcemap: sourcemapBytes,
-      bytecode: bytecodeBytes,
-      moduleInfo: moduleInfoBytes,
-      bytecodeOriginPath: bytecodeOriginPathBytes,
-      encoding: module.encoding,
-      loader: module.loader,
-      moduleFormat: module.moduleFormat,
-      side: module.side,
-    });
-
-    if (moduleStructSize === SIZEOF_MODULE_NEW) {
-      stringsData.push(
-        nameBytes,
-        contentsBytes,
-        sourcemapBytes,
-        bytecodeBytes,
-        moduleInfoBytes,
-        bytecodeOriginPathBytes
-      );
-    } else {
-      stringsData.push(nameBytes, contentsBytes, sourcemapBytes, bytecodeBytes);
-    }
-    return undefined;
-  });
-
-  const stringsPerModule = moduleStructSize === SIZEOF_MODULE_NEW ? 6 : 4;
-
-  // Phase 2: Calculate buffer layout
-  let currentOffset = 0;
-  const stringOffsets: StringPointer[] = [];
-
-  // Allocate space for strings with null terminators
-  for (const stringData of stringsData) {
-    stringOffsets.push({ offset: currentOffset, length: stringData.length });
-    currentOffset += stringData.length + 1; // +1 for null terminator
-  }
-
-  // Module structures
-  const modulesListOffset = currentOffset;
-  const modulesListSize = modulesMetadata.length * moduleStructSize;
-  currentOffset += modulesListSize;
-
-  // compileExecArgv
-  const compileExecArgvBytes = getStringPointerContent(
-    bunData,
-    bunOffsets.compileExecArgvPtr
-  );
-  const compileExecArgvOffset = currentOffset;
-  const compileExecArgvLength = compileExecArgvBytes.length;
-  currentOffset += compileExecArgvLength + 1; // +1 for null terminator
-
-  // Offsets structure
-  const offsetsOffset = currentOffset;
-  currentOffset += SIZEOF_OFFSETS;
-
-  // Trailer
-  const trailerOffset = currentOffset;
-  currentOffset += BUN_TRAILER.length;
-
-  // Phase 3: Build the new buffer
-  const newBuffer = Buffer.allocUnsafe(currentOffset);
-  newBuffer.fill(0);
-
-  // Write all strings with null terminators
-  let stringIdx = 0;
-  for (const { offset, length } of stringOffsets) {
-    if (length > 0) {
-      stringsData[stringIdx].copy(newBuffer, offset, 0, length);
-    }
-    newBuffer[offset + length] = 0; // null terminator
-    stringIdx++;
-  }
-
-  // Write compileExecArgv
-  if (compileExecArgvLength > 0) {
-    compileExecArgvBytes.copy(
-      newBuffer,
-      compileExecArgvOffset,
-      0,
-      compileExecArgvLength
-    );
-    newBuffer[compileExecArgvOffset + compileExecArgvLength] = 0;
-  }
-
-  // Build and write module structures
-  for (let i = 0; i < modulesMetadata.length; i++) {
-    const metadata = modulesMetadata[i];
-    const baseStringIdx = i * stringsPerModule;
-
-    const moduleStruct: BunModule = {
-      name: stringOffsets[baseStringIdx],
-      contents: stringOffsets[baseStringIdx + 1],
-      sourcemap: stringOffsets[baseStringIdx + 2],
-      bytecode: stringOffsets[baseStringIdx + 3],
-      moduleInfo:
-        moduleStructSize === SIZEOF_MODULE_NEW
-          ? stringOffsets[baseStringIdx + 4]
-          : { offset: 0, length: 0 },
-      bytecodeOriginPath:
-        moduleStructSize === SIZEOF_MODULE_NEW
-          ? stringOffsets[baseStringIdx + 5]
-          : { offset: 0, length: 0 },
-      encoding: metadata.encoding,
-      loader: metadata.loader,
-      moduleFormat: metadata.moduleFormat,
-      side: metadata.side,
-    };
-
-    // Serialize module structure inline
-    const moduleOffset = modulesListOffset + i * moduleStructSize;
-    let pos = moduleOffset;
-
-    // Write StringPointers (common to both formats)
-    newBuffer.writeUInt32LE(moduleStruct.name.offset, pos);
-    newBuffer.writeUInt32LE(moduleStruct.name.length, pos + 4);
-    pos += 8;
-    newBuffer.writeUInt32LE(moduleStruct.contents.offset, pos);
-    newBuffer.writeUInt32LE(moduleStruct.contents.length, pos + 4);
-    pos += 8;
-    newBuffer.writeUInt32LE(moduleStruct.sourcemap.offset, pos);
-    newBuffer.writeUInt32LE(moduleStruct.sourcemap.length, pos + 4);
-    pos += 8;
-    newBuffer.writeUInt32LE(moduleStruct.bytecode.offset, pos);
-    newBuffer.writeUInt32LE(moduleStruct.bytecode.length, pos + 4);
-    pos += 8;
-
-    // Write new-format-only StringPointers
-    if (moduleStructSize === SIZEOF_MODULE_NEW) {
-      newBuffer.writeUInt32LE(moduleStruct.moduleInfo.offset, pos);
-      newBuffer.writeUInt32LE(moduleStruct.moduleInfo.length, pos + 4);
-      pos += 8;
-      newBuffer.writeUInt32LE(moduleStruct.bytecodeOriginPath.offset, pos);
-      newBuffer.writeUInt32LE(moduleStruct.bytecodeOriginPath.length, pos + 4);
-      pos += 8;
-    }
-
-    // Write enum fields
-    newBuffer.writeUInt8(moduleStruct.encoding, pos);
-    newBuffer.writeUInt8(moduleStruct.loader, pos + 1);
-    newBuffer.writeUInt8(moduleStruct.moduleFormat, pos + 2);
-    newBuffer.writeUInt8(moduleStruct.side, pos + 3);
-  }
-
-  // Build and write Offsets structure inline
-  const newOffsets: BunOffsets = {
-    byteCount: offsetsOffset,
-    modulesPtr: {
-      offset: modulesListOffset,
-      length: modulesListSize,
-    },
-    entryPointId: bunOffsets.entryPointId,
-    compileExecArgvPtr: {
-      offset: compileExecArgvOffset,
-      length: compileExecArgvLength,
-    },
-    flags: bunOffsets.flags,
-  };
-
-  let offsetsPos = offsetsOffset;
-  const byteCount =
-    typeof newOffsets.byteCount === 'bigint'
-      ? newOffsets.byteCount
-      : BigInt(newOffsets.byteCount);
-  newBuffer.writeBigUInt64LE(byteCount, offsetsPos);
-  offsetsPos += 8;
-  newBuffer.writeUInt32LE(newOffsets.modulesPtr.offset, offsetsPos);
-  newBuffer.writeUInt32LE(newOffsets.modulesPtr.length, offsetsPos + 4);
-  offsetsPos += 8;
-  newBuffer.writeUInt32LE(newOffsets.entryPointId, offsetsPos);
-  offsetsPos += 4;
-  newBuffer.writeUInt32LE(newOffsets.compileExecArgvPtr.offset, offsetsPos);
-  newBuffer.writeUInt32LE(newOffsets.compileExecArgvPtr.length, offsetsPos + 4);
-  offsetsPos += 8;
-  newBuffer.writeUInt32LE(newOffsets.flags, offsetsPos);
-
-  // Write trailer
-  BUN_TRAILER.copy(newBuffer, trailerOffset);
-
-  return newBuffer;
-}
-
 /**
  * Atomically writes a binary using LIEF and copies permissions from original.
  * Includes robust handling for busy/executing files.
@@ -1064,24 +424,6 @@ function atomicWriteBinary(
 
     throw error;
   }
-}
-
-/**
- * Builds section data with size header followed by content.
- * Format: [size header][content]
- *
- * @param bunBuffer - The bun data buffer to wrap
- * @param headerSize - Header size: 4 for old format (Bun < 1.3.4), 8 for new format (default)
- */
-function buildSectionData(bunBuffer: Buffer, headerSize: number = 8): Buffer {
-  const sectionData = Buffer.allocUnsafe(headerSize + bunBuffer.length);
-  if (headerSize === 8) {
-    sectionData.writeBigUInt64LE(BigInt(bunBuffer.length), 0);
-  } else {
-    sectionData.writeUInt32LE(bunBuffer.length, 0);
-  }
-  bunBuffer.copy(sectionData, headerSize);
-  return sectionData;
 }
 
 function repackMachO(
@@ -1257,78 +599,6 @@ function repackELF(
     console.error('repackELF failed:', error);
     throw error;
   }
-}
-
-/** Repacks a modified Bun buffer into an ELF binary by rewriting the file overlay. */
-function repackELFRaw(
-  binPath: string,
-  newBunBuffer: Buffer,
-  outputPath: string
-): void {
-  const originalData = fs.readFileSync(binPath);
-  const fileSize = originalData.length;
-
-  const tailSize = SIZEOF_OFFSETS + BUN_TRAILER.length + 8;
-  if (fileSize < tailSize) {
-    throw new Error('repackELFRaw: file too small to contain Bun data');
-  }
-  const offsetsBytes = originalData.subarray(
-    fileSize - tailSize,
-    fileSize - tailSize + SIZEOF_OFFSETS
-  );
-  const existingOffsets = parseOffsets(offsetsBytes);
-  const byteCount =
-    typeof existingOffsets.byteCount === 'bigint'
-      ? Number(existingOffsets.byteCount)
-      : existingOffsets.byteCount;
-
-  const overlaySize = byteCount + SIZEOF_OFFSETS + BUN_TRAILER.length + 8;
-  const elfSize = fileSize - overlaySize;
-
-  if (elfSize <= 0 || elfSize >= fileSize) {
-    throw new Error(`repackELFRaw: computed ELF size out of range: ${elfSize}`);
-  }
-
-  debug(
-    `repackELFRaw: elfSize=${elfSize}, byteCount=${byteCount}, newBunBuffer=${newBunBuffer.length}`
-  );
-
-  const newOverlay = Buffer.allocUnsafe(newBunBuffer.length + 8);
-  newBunBuffer.copy(newOverlay, 0);
-  newOverlay.writeBigUInt64LE(BigInt(newBunBuffer.length), newBunBuffer.length);
-
-  const newBinary = Buffer.concat([
-    originalData.subarray(0, elfSize),
-    newOverlay,
-  ]);
-
-  const tempPath = outputPath + '.tmp';
-  fs.writeFileSync(tempPath, newBinary);
-  const origStat = fs.statSync(binPath);
-  fs.chmodSync(tempPath, origStat.mode);
-  try {
-    fs.renameSync(tempPath, outputPath);
-  } catch (error) {
-    try {
-      fs.unlinkSync(tempPath);
-    } catch {
-      // ignore cleanup errors
-    }
-    if (
-      error instanceof Error &&
-      'code' in error &&
-      (error.code === 'ETXTBSY' ||
-        error.code === 'EBUSY' ||
-        error.code === 'EPERM')
-    ) {
-      throw new Error(
-        'Cannot update the Claude executable while it is running.\n' +
-          'Please close all Claude instances and try again.'
-      );
-    }
-    throw error;
-  }
-  debug('repackELFRaw: Write completed successfully');
 }
 
 /**
