+++ src/tests/config.test.ts
@@ -1524,90 +1524,6 @@ describe('config.ts', () => {
         findClaudeCodeInstallation(mockConfig, { interactive: true })
       ).rejects.toThrow('Could not extract JS from native binary');
     });
-
-    it('should include node-lief load reason in error when getNativeModuleLoadError returns a message', async () => {
-      const mockConfig = {
-        ccInstallationPath: null,
-        changesApplied: false,
-        ccVersion: '',
-        lastModified: '',
-        settings: DEFAULT_SETTINGS,
-      };
-
-      const nativeBinaryPath = '/usr/local/bin/claude';
-
-      vi.spyOn(fs, 'stat').mockImplementation(async filePath => {
-        if (filePath === nativeBinaryPath) return {} as Stats;
-        throw createEnoent();
-      });
-      vi.mocked(whichMock).mockResolvedValue(nativeBinaryPath);
-      lstatSpy.mockImplementation(async filePath => {
-        if (filePath === nativeBinaryPath) return createRegularStats();
-        throw createEnoent();
-      });
-      vi.spyOn(fs, 'realpath').mockResolvedValue(nativeBinaryPath);
-      vi.spyOn(fs, 'open').mockResolvedValue({
-        read: async ({ buffer }: { buffer: Buffer }) => {
-          const contentBuffer = Buffer.from('fake binary content');
-          contentBuffer.copy(buffer);
-          return { bytesRead: contentBuffer.length, buffer };
-        },
-        close: async () => {},
-      } as unknown as fs.FileHandle);
-      mockMagicInstance.detect.mockReturnValue('application/octet-stream');
-      vi.mocked(
-        nativeInstallation.extractClaudeJsFromNativeInstallation
-      ).mockResolvedValue(null);
-      vi.mocked(nativeInstallation.getNativeModuleLoadError).mockReturnValue(
-        'libstdc++.so.6: cannot open shared object file: No such file or directory'
-      );
-
-      await expect(
-        findClaudeCodeInstallation(mockConfig, { interactive: true })
-      ).rejects.toThrow('node-lief failed to load');
-    });
-
-    it('should include NixOS hint in error when load error contains shared object message', async () => {
-      const mockConfig = {
-        ccInstallationPath: null,
-        changesApplied: false,
-        ccVersion: '',
-        lastModified: '',
-        settings: DEFAULT_SETTINGS,
-      };
-
-      const nativeBinaryPath = '/usr/local/bin/claude';
-
-      vi.spyOn(fs, 'stat').mockImplementation(async filePath => {
-        if (filePath === nativeBinaryPath) return {} as Stats;
-        throw createEnoent();
-      });
-      vi.mocked(whichMock).mockResolvedValue(nativeBinaryPath);
-      lstatSpy.mockImplementation(async filePath => {
-        if (filePath === nativeBinaryPath) return createRegularStats();
-        throw createEnoent();
-      });
-      vi.spyOn(fs, 'realpath').mockResolvedValue(nativeBinaryPath);
-      vi.spyOn(fs, 'open').mockResolvedValue({
-        read: async ({ buffer }: { buffer: Buffer }) => {
-          const contentBuffer = Buffer.from('fake binary content');
-          contentBuffer.copy(buffer);
-          return { bytesRead: contentBuffer.length, buffer };
-        },
-        close: async () => {},
-      } as unknown as fs.FileHandle);
-      mockMagicInstance.detect.mockReturnValue('application/octet-stream');
-      vi.mocked(
-        nativeInstallation.extractClaudeJsFromNativeInstallation
-      ).mockResolvedValue(null);
-      vi.mocked(nativeInstallation.getNativeModuleLoadError).mockReturnValue(
-        'libstdc++.so.6: cannot open shared object file: No such file or directory'
-      );
-
-      await expect(
-        findClaudeCodeInstallation(mockConfig, { interactive: true })
-      ).rejects.toThrow('On NixOS with Bun');
-    });
   });
 
   describe('startupCheck', () => {
